Fixed bad refactoring
