+++ library/src/scripts/flyouts/items/DropDownItemButton.tsx
@@ -6,7 +6,7 @@
 
 import React from "react";
 import Button from "@library/forms/Button";
-import { ButtonTypes } from "@library/forms/buttonTypes";
+import { ButtonTypes } from "@library/forms/buttonStyles";
 import { dropDownClasses } from "@library/flyouts/dropDownStyles";
 import { ISelectBoxItem } from "@library/forms/select/SelectBox";
 import classNames from "classnames";
