+++ library/src/scripts/forms/select/ClearButton.tsx
@@ -7,7 +7,7 @@
 import React from "react";
 import { t } from "@library/utility/appUtils";
 import Button from "@library/forms/Button";
-import { ButtonTypes } from "@library/forms/buttonTypes";
+import { ButtonTypes } from "@library/forms/buttonStyles";
 import { clear } from "@library/icons/common";
 import classNames from "classnames";
