+++ library/src/scripts/headers/mebox/pieces/MessagesContents.tsx
@@ -10,7 +10,7 @@ import FrameHeaderWithAction from "@library/layout/frame/FrameHeaderWithAction";
 import { IConversationsStoreState } from "@library/features/conversations/ConversationsModel";
 import Loader from "@library/loaders/Loader";
 import { loaderClasses } from "@library/loaders/loaderStyles";
-import { ButtonTypes, buttonUtilityClasses } from "@library/forms/buttonTypes";
+import { ButtonTypes, buttonUtilityClasses } from "@library/forms/buttonStyles";
 import apiv2 from "@library/apiv2";
 import { withDevice, Devices, IDeviceProps } from "@library/layout/DeviceContext";
 import { IMeBoxMessageItem, MeBoxItemType } from "@library/headers/mebox/pieces/MeBoxDropDownItem";
