+++ library/src/scripts/navigation/CloseButton.tsx
@@ -6,7 +6,7 @@
 import React from "react";
 import classNames from "classnames";
 import { t } from "@library/utility/appUtils";
-import { ButtonTypes } from "@library/forms/buttonTypes";
+import { ButtonTypes } from "@library/forms/buttonStyles";
 import Button from "@library/forms/Button";
 import { close, closeCompact } from "@library/icons/common";
