+++ library/src/scripts/navigation/simplePagerStyles.ts
@@ -8,7 +8,7 @@ import { unit } from "@library/styles/styleHelpers";
 import { styleFactory, useThemeCache, variableFactory } from "@library/styles/styleUtils";
 import { globalVariables } from "@library/styles/globalStyleVars";
 import { formElementsVariables } from "@library/forms/formElementStyles";
-import { buttonGlobalVariables } from "@library/forms/buttonTypes";
+import { buttonGlobalVariables } from "@library/forms/buttonStyles";
 
 export const simplePagerVariables = useThemeCache(() => {
     const themeVars = variableFactory("simplePage");
