+++ plugins/rich-editor/src/scripts/flyouts/EmbedFlyout.tsx
@@ -6,7 +6,7 @@
 
 import DropDown, { FlyoutSizes } from "@library/flyouts/DropDown";
 import Button from "@library/forms/Button";
-import { ButtonTypes } from "@library/forms/buttonTypes";
+import { ButtonTypes } from "@library/forms/buttonStyles";
 import { embed } from "@library/icons/editorIcons";
 import Frame from "@library/layout/frame/Frame";
 import FrameBody from "@library/layout/frame/FrameBody";
