+++ plugins/rich-editor/src/scripts/flyouts/EmojiFlyout.tsx
@@ -18,8 +18,8 @@ import classNames from "classnames";
 import EmojiPicker from "@rich-editor/flyouts/pieces/EmojiPicker";
 import { richEditorClasses } from "@rich-editor/editor/richEditorClasses";
 import { forceSelectionUpdate } from "@rich-editor/quill/utility";
-import { ButtonTypes } from "@library/forms/buttonTypes";
 import { IconForButtonWrap } from "@rich-editor/editor/pieces/IconForButtonWrap";
+import { ButtonTypes } from "@library/forms/buttonStyles";
 
 interface IProps {
     disabled?: boolean;
