+++ plugins/rich-editor/src/scripts/flyouts/pieces/Flyout.tsx
@@ -12,7 +12,7 @@ import { IWithEditorProps } from "@rich-editor/editor/context";
 import { withEditor } from "@rich-editor/editor/withEditor";
 import { richEditorFlyoutClasses } from "@rich-editor/flyouts/pieces/flyoutClasses";
 import { richEditorClasses } from "@rich-editor/editor/richEditorClasses";
-import { ButtonTypes } from "@library/forms/buttonTypes";
+import { ButtonTypes } from "@library/forms/buttonStyles";
 
 interface IState {
     id: string;
